client/requirements.txt
python-dotenv
psutil